requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
anthropic==0.58.2
dynaconf==3.2.4
numpy==1.24.3
httpx==0.27.2